pyproject.toml
[project]
name = "Synnovator"
version = "0.1.0"
description = "Synnovator is a platform for creating and managing your own AI-powered hackathon."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=5.2.10",
    "django-storages[s3]>=1.14.6",
    "gunicorn==23.0.0",
    "psycopg[binary]>=3.3.2",
    "wagtail>=7.2.1",
    "wagtail-localize>=1.10",
    "wagtail-storages>=2.0",
    "whitenoise>=6.11.0",
    "polib>=1.2.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "django-debug-toolbar>=5.0.0",
    "faker>=33.1.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "wagtail-factories>=4.1.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "synnovator.settings.dev"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "wagtail: marks tests requiring Wagtail setup",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["synnovator"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/management/commands/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
fail_under = 60
